Setup for pentest-env
Change for supporting custom box as the base windows box to use to
provision the target. Tested with this box:
https://atlas.hashicorp.com/opentable/boxes/win-2008r2-standard-amd64-nocm

- Set GUI to true (for debugging)
- Set memory to 4096
- Start provisioning by applying password policy. After this
provisioning script, vagrant username will be renamed to
'Administrator'. A permission error will occure, fix this by
uncommenting 'winrm.username' config.
- Install root certs with shell provisioning (They are currently
installed with packer provisioning when building the windows2008 box)
- Fixed some path due to change of administrator username.